Make the ClusterRoleBinding in templates dir a template
Currently the ClusterRoleBinding in the templates dir binds the
ServiceAccount in the namespace `default` to the `backupjob`
ClusterRole. However, the ServiceAccount may be in a different
namespace. This change makes the ServiceAccount namespace be able
to be overridden using a parameter.

Our ServiceAccount, ClusterRole and ClusterRoleBinding all are
hardcoded to use the `backupjob` names and to reference each other
using that hardcoded name. This doesn't seem to be an issue at the
moment as each namespace will only have one ServiceAccount that
all backup Jobs and CronJobs in that namespace will use. So the
names have not been parameterized.